+++ Strava_Client/src/es/deusto/ingenieria/sd/strava/client/controller/AthleteController.java
@@ -15,7 +15,7 @@ public class AthleteController {
         this.serviceLocator = serviceLocator;
     }
 
-    public boolean register(String email, String password, String name, Date birthDate, Float weight, Integer height,
+    public boolean register(String email, String password, String name, Date birthDate, Double weight, Integer height,
             Integer restingHeartrate, Integer maxHeartrate) {
         try {
             this.token = this.serviceLocator.getService().register(email, password, name, birthDate, weight, height, restingHeartrate, maxHeartrate);
@@ -26,7 +26,7 @@ public class AthleteController {
         }
     }
 
-    public boolean registerGoogle(String email, String password, String name, Date birthDate, Float weight, Integer height,
+    public boolean registerGoogle(String email, String password, String name, Date birthDate, Double weight, Integer height,
             Integer restingHeartrate, Integer maxHeartrate) {
         try {
             this.token = this.serviceLocator.getService().registerGoogle(email, password, name, birthDate, weight, height, restingHeartrate, maxHeartrate);
@@ -37,7 +37,7 @@ public class AthleteController {
         }
     }
 
-    public boolean registerFacebook(String email, String password, String name, Date birthDate, Float weight, Integer height,
+    public boolean registerFacebook(String email, String password, String name, Date birthDate, Double weight, Integer height,
             Integer restingHeartrate, Integer maxHeartrate) {
         try {
             this.token = this.serviceLocator.getService().registerFacebook(email, password, name, birthDate, weight, height, restingHeartrate, maxHeartrate);
